refactor: wydziel usePendingOperations z App.jsx
Nowy hook src/hooks/usePendingOperations.js hermetyzuje:
- pendingCount, isSyncing, wasOfflineRef
- effect sync kolejki offline po powrocie online
- refreshPendingCount() po każdym add/update/delete

Dodatkowo: onSynced i addToast trzymane w refach, więc zmiana ich
identyczności nie re-triggeruje effectu sync. Poprzednia inline wersja
miała fetchData w depsach, co powodowało uruchamianie effectu przy
każdej zmianie currentPeriod (z guardem, więc bezpieczne, ale zbędne).

App.jsx: usunięty import getQueuedOperations/processQueue.
Trzy call-site setPendingCount(getQueuedOperations().length) → refreshPendingCount().

Testy: 205/205. Build OK.
